천운-app/palm-api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
opencv-python-headless==4.9.0.80
mediapipe==0.10.9
numpy==1.26.3
Pillow==10.2.0
scikit-image==0.22.0
